refactor(electron): main.ts 拆 4 module + 收口 (主题 E #62 C-303+C-402)
main.ts 原 388 行 god 拆为:
- main-debug.ts (4 KB): debug 模式 + DevTools + AUMID + app icon
- main-splash.ts (3.8 KB): splash 窗口生命周期 (create / fade / destroy / timeout)
- main-protocol.ts (2 KB): x-agent-logos:// 自定义协议
- main-nav-guard.ts (2.3 KB): 外部 URL 导航守卫

main.ts 388 -> 137 行 (-251, -65%). 仍保留:
- app.whenReady 编排 + second-instance / activate / window-all-closed 处理
- 4 module 调用 (createDebugShortcuts / createRevealMain / registerLogoProtocolHandler / etc)
- 单实例锁 (X_AGENT_ALLOW_MULTI)

不在本 PR:
- app-runtime.ts:492 行的 registerIpc 函数 7 类 inline 收口 (plan #62 C-402):
  - 计划说"已被 8 个 register-*.ts 替代", 实际只有 workspace/turn/plan/session-config/
    provider/godot/update 7 个模块化, 还剩 ~15 个 inline handle() (openProject /
    getPrefs / setPrefs / checkBash / listProjectDir / readProjectFile / listPlugins
    / listInstalledPackages / installPackage / uninstallPackage / openPiLogin 等)
  - 全部移到 register-*.ts 涉及更多新文件 + 巨大 PR 体积, 留 0.7.x
  - 触发重新评估条件: 拆后接口 ≈ 子集之和 → 留 ADR 不追拆

收益:
- main.ts 388 -> 137 行 (-65%)
- 4 module 各自 < 4KB, 单一关注点
- typecheck 0 errors, vitest 543 pass (无新 case, 纯重构)
- 启动流程 4 段 (debug / splash / protocol / nav-guard) 各自 < 4KB, 易于单测

验收:
- npm run typecheck: 0 errors
- npx vitest run: 41 files / 543 tests pass
- main.ts < 200 行 (实际 137)
- 4 module 各自 vitest 可独立 mock (caller 用 vi.fn 替代即可)

Refs #62 (主题 E) - main.ts 拆分完成; app-runtime.ts IPC inline 收口留 0.7.x
